custom hook for image handle and form validation
