+++ my-blog/src/index.css
@@ -101,6 +101,9 @@ main {
   margin: 2rem 0.5rem;
   padding: 0.25rem
 }
+.error-message {
+  color: red;
+}
 
 @media (prefers-color-scheme: light) {
   :root {
